update typer imports
